# api/openapi.yaml
openapi: 3.1.0
info:
  title: AppLab API
  description: |-
    AppLab deploys an application to Kubernetes by uploading its source.

    Every JSON route answers inside a `{"data": ...}` envelope, except `/health`; failures answer `{"error": "...", "retryable": bool}`. Three routes answer `text/plain` and may stream (`x-streaming`), which a client must read incrementally rather than buffer.

    This document is generated from AppLab's route table. Do not edit it by hand — run `make gen-openapi`.
  version: v1
servers:
  - url: '{baseUrl}'
    description: The deployment's address, including any base path (for example https://applab.example.com/applab).
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/apps/{app}/builds:
    get:
      operationId: listBuilds
      summary: An app's builds, newest first, read from the build Jobs.
      description: An app's builds, newest first, read from the build Jobs. `?limit=` (default 20). How far back this reaches is build.ttlAfterFinished, because the Job is the record.
      tags:
        - builds
      parameters:
        - name: app
          in: path
          required: true
          schema:
            type: string
          example: shop
        - name: limit
          in: query
          description: Maximum number of items to return.
          schema:
            type: integer
          example: 20
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        app_id:
                          type: string
                        branch:
                          type: string
                        commit_sha:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        id:
                          type: string
                        image:
                          type: string
                        job_name:
                          type: string
                        pod:
                          $ref: '#/components/schemas/Pod'
                          nullable: true
                        reason:
                          type: string
                        status:
                          type: string
                      required:
                        - app_id
                        - branch
                        - commit_sha
                        - created_at
                        - id
                        - image
                        - job_name
                        - pod
                        - reason
                        - status
                required:
                  - data
        default:
          description: Failure. `retryable` says whether the same call could succeed on a retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-applab-tier: app
    post:
      operationId: startBuild
      summary: Build an image from a commit.
      description: Build an image from a commit. Body `{commit_sha?, deploy?}` — omit `commit_sha` to build the current tip. Returns immediately; the build runs as a Job in the cluster. A build deploys what it builds once it succeeds, the same way a push does, subject to the app's `auto_deploy` switch and to the branch being the app's active one; pass `"deploy":false` (or `?deploy=false`) to build the image and stop. Returns 501 if this deployment cannot build.
      tags:
        - builds
      parameters:
        - name: app
          in: path
          required: true
          schema:
            type: string
          example: shop
        - name: branch
          in: query
          description: Build another branch without switching to it.
          schema:
            type: string
        - name: deploy
          in: query
          description: Deploy the commit once the build succeeds. Defaults to true; send false to build the image and stop.
          schema:
            type: boolean
          example: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildRequest'
      responses:
        "202":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      app_id:
                        type: string
                      branch:
                        type: string
                      commit_sha:
                        type: string
                      created_at:
                        type: string
                        format: date-time
                      id:
                        type: string
                      image:
                        type: string
                      job_name:
                        type: string
                      pod:
                        $ref: '#/components/schemas/Pod'
                        nullable: true
                      reason:
                        type: string
                      status:
                        type: string
                    required:
                      - app_id
                      - branch
                      - commit_sha
                      - created_at
                      - id
                      - image
                      - job_name
                      - pod
                      - reason
                      - status
                required:
                  - data
        default:
          description: Failure. `retryable` says whether the same call could succeed on a retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-applab-tier: app
...
components:
  schemas:
...
    BuildRequest:
      type: object
      properties:
        branch:
          type: string
        commit_sha:
          type: string
        deploy:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
        retryable:
          type: boolean
...
    Pod:
      type: object
      properties:
        containers:
          type: array
          items:
            type: object
        image:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
        name:
          type: string
        phase:
          type: string
        ready:
          type: boolean
        reason:
          type: string
        restarts:
          type: integer
          format: int32
      required:
        - containers
        - image
        - labels
        - message
        - name
        - phase
        - ready
        - reason
        - restarts
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 'The git transport''s scheme only: the password is an API key and the username is ignored.'
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key as a bearer token. An admin key reaches everything; an app key reaches one app (see x-applab-tier on each operation).